requirements.txt
######### Here are the packages required to run the code #################

Pillow
ipywidgets